+++ components/camel-saxon/src/test/java/org/apache/camel/builder/saxon/XQueryHeaderNameResultTypeAndNamespaceTest.java
@@ -17,8 +17,8 @@
 package org.apache.camel.builder.saxon;
 
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.builder.xml.Namespaces;
 import org.apache.camel.component.mock.MockEndpoint;
+import org.apache.camel.support.builder.xml.Namespaces;
 import org.apache.camel.test.junit4.CamelTestSupport;
 import org.junit.Test;
