+++ tests/camel-itest/src/test/java/org/apache/camel/itest/issues/RssToHttpIssueTest.java
@@ -18,8 +18,8 @@ package org.apache.camel.itest.issues;
 
 import org.apache.camel.Exchange;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.builder.xml.Namespaces;
 import org.apache.camel.component.mock.MockEndpoint;
+import org.apache.camel.support.builder.xml.Namespaces;
 import org.apache.camel.test.junit4.CamelTestSupport;
 import org.junit.Ignore;
 import org.junit.Test;
